+++ packages/horizon-storage/src/horizon_storage/exceptions/__init__.py
@@ -1,0 +1,5 @@
+"""Storage exceptions."""
+
+from horizon_storage.exceptions.storage import StorageCorruptionError, StorageError
+
+__all__ = ["StorageCorruptionError", "StorageError"]